cli: reorganize code
Subcommands now go into their own file.
