+++ shell.nix
@@ -18,7 +18,7 @@ mkShell {
   # fortify does not work with -O0 which is used by spdk when --enable-debug
   hardeningDisable = [ "fortify" ];
   buildInputs = [
-    channel.nightly.rust
+    #    channel.nightly.rust
     clang
     cowsay
     e2fsprogs
